// src/domain/types/actor/parts/SocialSkills.ts
import { FoundryCalculableField } from "../fields/FoundryCalculableField";

export type SocialSkills = {
  style: FoundryCalculableField;
  intimidate: FoundryCalculableField;
  leadership: FoundryCalculableField;
  persuasion: FoundryCalculableField;
  trading: FoundryCalculableField;
  streetwise: FoundryCalculableField;
  etiquette: FoundryCalculableField;
};