// src/domain/types/actor/parts/CreativeSkills.ts
import { FoundryCalculableField } from "../fields/FoundryCalculableField";

export type CreativeSkills = {
  art: FoundryCalculableField;
  dance: FoundryCalculableField;
  forging: FoundryCalculableField;
  runes: FoundryCalculableField;
  alchemy: FoundryCalculableField;
  animism: FoundryCalculableField;
  music: FoundryCalculableField;
  sleightOfHand: FoundryCalculableField;
  ritualCalligraphy: FoundryCalculableField;
  jewelry: FoundryCalculableField;
  tailoring: FoundryCalculableField;
  puppetMaking: FoundryCalculableField;
};