// src/domain/types/actor/parts/IntellectualSkills.ts
import { FoundryCalculableField } from "../fields/FoundryCalculableField";

export type IntellectualSkills = {
  animals: FoundryCalculableField;
  science: FoundryCalculableField;
  law: FoundryCalculableField;
  herbalLore: FoundryCalculableField;
  history: FoundryCalculableField;
  tactics: FoundryCalculableField;
  medicine: FoundryCalculableField;
  memorize: FoundryCalculableField;
  navigation: FoundryCalculableField;
  occult: FoundryCalculableField;
  appraisal: FoundryCalculableField;
  magicAppraisal: FoundryCalculableField;
};