// src/domain/types/actor/parts/SubterfugeSkills.ts
import { FoundryCalculableField } from "../fields/FoundryCalculableField";

export type SubterfugeSkills = {
  lockPicking: FoundryCalculableField;
  disguise: FoundryCalculableField;
  hide: FoundryCalculableField;
  theft: FoundryCalculableField;
  stealth: FoundryCalculableField;
  trapLore: FoundryCalculableField;
  poisons: FoundryCalculableField;
};